pyproject.toml
[project]
name = "genai-document-portal"
version = "0.1.0"
description = "End to end LLMOps project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.129.0",
    "from-root>=1.3.0",
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langchain-mongodb>=0.11.0",
    "langchain-text-splitters>=1.1.1",
    "langsmith>=0.7.4",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]
